test(version): split test_version_pep440 into test_segments and test_version, move tests for "next"-methods from semver/test_version to pep440/test_version
